+++ Gui/WindowFrames/FilterButton/FilterButtonMixin.lua
@@ -2,8 +2,12 @@ local _, addon = ...;
 
 KrowiAF_AchievementFrameFilterButtonMixin = {};
 
+local LoadNew;
 function KrowiAF_AchievementFrameFilterButtonMixin:OnLoad()
 	self:SetFrameLevel(self:GetParent():GetFrameLevel() + 7);
+    if addon.Util.IsTheWarWithin then
+        LoadNew();
+    end
 end
 
 function KrowiAF_AchievementFrameFilterButtonMixin:OnShow()
@@ -310,12 +314,6 @@ function KrowiAF_AchievementFrameFilterButtonMixin:OnMouseDown()
 
     AddAchievementFilters(menu, nil, addon.Filters.db.profile);
 
-    -- local version = addon.Objects.MenuItem:New(addon.L["Version"]);
-    -- for _, buildVersion in next, addon.Data.BuildVersions do
-    --     AddCheckBox(version, buildVersion.Name, addon.Filters.db.profile, {"BuildVersion", buildVersion.Id});
-    -- end
-    -- menu:Add(version);
-
     local earnedBy = addon.Objects.MenuItem:New(addon.L["Earned By"]);
     AddRadioButton(menu, earnedBy, addon.Filters.Account, addon.Filters.db.profile, {"EarnedBy"}, false);
     AddRadioButton(menu, earnedBy, addon.Filters.CharacterAccount, addon.Filters.db.profile, {"EarnedBy"}, false);
@@ -371,4 +369,232 @@ function KrowiAF_AchievementFrameFilterButtonMixin:OnMouseDown()
     end
 
 	menu:Toggle(self, 96, 15);
+end
+
+-- The War Within DropDown rework
+-- TODO Anchor + OnShow + OnHide
+if not addon.Util.IsTheWarWithin then
+    return;
+end
+
+local function CreateCheckbox(_menu, text, filters, keys, checkTabs)
+    _menu:CreateCheckbox(text,
+        function()
+            return addon.Util.ReadNestedKeys(filters, keys);
+        end,
+        function()
+            SetSelected(filters, keys, not addon.Util.ReadNestedKeys(filters, keys), checkTabs, true);
+            UpdateAchievementFrame();
+        end
+    );
+end
+
+local function CreateRadio(_menu, text, filters, keys, checkTabs)
+    _menu:CreateRadio(text,
+        function()
+            return addon.Util.ReadNestedKeys(filters, keys) == text;
+        end,
+        function()
+            SetSelected(filters, keys, text, checkTabs, false);
+            KrowiAF_AchievementsFrame:ForceUpdate();
+        end
+    );
+end
+
+local function IsMinorVersionChecked(filters, minor)
+    local isChecked = true;
+    for _, patch in next, minor.Patches do
+        isChecked = isChecked and addon.Util.ReadNestedKeys(filters, {"BuildVersion", patch.BuildVersionId});
+    end
+    return isChecked;
+end
+
+local function CreateMinorVersion(majorGroup, filters, major, minor)
+    if #minor.Patches <= 1 then
+        CreateCheckbox(majorGroup, major.Major .. "." .. minor.Minor .. "." .. minor.Patches[1].Patch, filters, {"BuildVersion", minor.Patches[1].BuildVersionId}, true);
+        return;
+    end
+    local minorGroup = majorGroup:CreateCheckbox(major.Major .. "." .. minor.Minor .. ".x",
+        function()
+            return IsMinorVersionChecked(filters, minor);
+        end,
+         function()
+            local checked = IsMinorVersionChecked(filters, minor);
+            for _, patch in next, minor.Patches do
+                SetSelected(filters, {"BuildVersion", patch.BuildVersionId}, not checked, true, true);
+            end
+            UpdateAchievementFrame();
+        end
+    );
+    for _, patch in next, minor.Patches do
+        CreateCheckbox(minorGroup, major.Major .. "." .. minor.Minor .. "." .. patch.Patch, filters, {"BuildVersion", patch.BuildVersionId}, true);
+    end
+end
+
+local function IsMajorVersionChecked(filters, major)
+    local isChecked = true;
+    for _, minor in next, major.Minors do
+        for _, patch in next, minor.Patches do
+            isChecked = isChecked and addon.Util.ReadNestedKeys(filters, {"BuildVersion", patch.BuildVersionId});
+        end
+    end
+    return isChecked;
+end
+
+local function CreateMajorVersion(version, filters, major)
+    local majorGroup = version:CreateCheckbox(major.Major .. ".x.x",
+        function()
+            return IsMajorVersionChecked(filters, major);
+        end,
+        function()
+            local checked = IsMajorVersionChecked(filters, major);
+            for _, minor in next, major.Minors do
+                for _, patch in next, minor.Patches do
+                    SetSelected(filters, {"BuildVersion", patch.BuildVersionId}, not checked, true, true);
+                end
+            end
+            UpdateAchievementFrame();
+        end);
+    for _, minor in next, major.Minors do
+        CreateMinorVersion(majorGroup, filters, major, minor);
+    end
+end
+
+local function CreateSelectDeselectAllVersions(version, filters)
+    version:CreateDivider();
+    version:CreateButton(addon.L["Select All"],
+        function()
+            for _, major in next, addon.Data.BuildVersionsGrouped do
+                for _, minor in next, major.Minors do
+                    for _, patch in next, minor.Patches do
+                        SetSelected(filters, {"BuildVersion", patch.BuildVersionId}, true, true, true);
+                    end
+                end
+            end
+            UpdateAchievementFrame();
+        end);
+    version:CreateButton(addon.L["Deselect All"],
+        function()
+            for _, major in next, addon.Data.BuildVersionsGrouped do
+                for _, minor in next, major.Minors do
+                    for _, patch in next, minor.Patches do
+                        SetSelected(filters, {"BuildVersion", patch.BuildVersionId}, false, true, true);
+                    end
+                end
+            end
+            UpdateAchievementFrame();
+        end);
+end
+
+local function CreateBuildVersionFilter(_menu, filters)
+    local version = _menu:CreateButton(addon.L["Version"]);
+    if #addon.Data.BuildVersionsGrouped > 1 then
+        for _, major in next, addon.Data.BuildVersionsGrouped do
+            CreateMajorVersion(version, filters, major)
+        end
+    else
+        local major = addon.Data.BuildVersionsGrouped[1];
+        for _, minor in next, major.Minors do
+            CreateMinorVersion(version, filters, major, minor);
+        end
+    end
+    CreateSelectDeselectAllVersions(version, filters);
+end
+
+local function CreateAchievementFilters(_menu, filters)
+    CreateCheckbox(_menu, addon.L["Completed"], filters, {"Completion", "Completed"}, true);
+    CreateCheckbox(_menu, addon.L["Not Completed"], filters, {"Completion", "NotCompleted"}, true);
+    CreateCheckbox(_menu, addon.L["Past Obtainable"], filters, {"Obtainability", "PastObtainable"}, true);
+    CreateCheckbox(_menu, addon.L["Current Obtainable"], filters, {"Obtainability", "CurrentObtainable"}, true);
+    CreateCheckbox(_menu, addon.L["Future Obtainable"], filters, {"Obtainability", "FutureObtainable"}, true);
+
+    CreateBuildVersionFilter(_menu, filters);
+
+    local faction = _menu:CreateButton(addon.L["Faction"]);
+    CreateCheckbox(faction, addon.L["Neutral"], filters, {"Faction", "Neutral"}, true);
+    CreateCheckbox(faction, addon.L["Alliance"], filters, {"Faction", "Alliance"}, true);
+    CreateCheckbox(faction, addon.L["Horde"], filters, {"Faction", "Horde"}, true);
+    faction:CreateDivider();
+    local function SetFactionFilters(value)
+        for _faction, _ in next, filters.Faction do
+            SetSelected(filters, {"Faction", _faction}, value, true, true);
+        end
+        UpdateAchievementFrame();
+    end
+    faction:CreateButton(addon.L["Select All"], SetFactionFilters, true);
+    faction:CreateButton(addon.L["Deselect All"], SetFactionFilters, false);
+
+    CreateCheckbox(_menu, addon.L["Realm First!"], filters, {"Special", "RealmFirst"}, true);
+    CreateCheckbox(_menu, addon.GetCategoryInfoTitle(81), filters, {"Special", "FeatsOfStrength"}, true);
+    CreateCheckbox(_menu, addon.GetCategoryInfoTitle(95), filters, {"Special", "PvP"}, true);
+    CreateCheckbox(_menu, addon.L["Collapse Chain"], filters, {"CollapseSeries"}, true);
+    if addon.Options.db.profile.Categories.TrackingAchievements.DoLoad then
+        CreateCheckbox(_menu, addon.L["Tracking"], filters, {"Tracking"}, true);
+    end
+    CreateCheckbox(_menu, addon.L["Excluded"], filters, {"Excluded"}, true);
+    CreateCheckbox(_menu, addon.L["Always Show Completed"], filters, {"Completion", "AlwaysShowCompleted"}, true);
+
+    _menu:CreateDivider();
+    local text = addon.L["Sort By"];
+    if filters == addon.Filters.db.profile and ValueIsUndecided({"SortBy", "Criteria"}) then
+        text = text .. " (*)";
+    end
+    local sortBy = _menu:CreateButton(text);
+    CreateRadio(sortBy, addon.L["Default"], filters, {"SortBy", "Criteria"}, true);
+    CreateRadio(sortBy, addon.L["Name"], filters, {"SortBy", "Criteria"}, true);
+    CreateRadio(sortBy, addon.L["Completion"], filters, {"SortBy", "Criteria"}, true);
+    CreateRadio(sortBy, addon.L["ID"], filters, {"SortBy", "Criteria"}, true);
+    CreateRadio(sortBy, addon.L["Points"], filters, {"SortBy", "Criteria"}, true);
+    sortBy:CreateDivider();
+    CreateCheckbox(sortBy, addon.L["Reverse Sort"], filters, {"SortBy", "ReverseSort"}, true, true);
+end
+
+function LoadNew()
+    KrowiAF_AchievementFrameFilterDropdown:SetupMenu(function(dropdown, rootDescription)
+		rootDescription:SetTag("MENU_ACHIEVEMENT_FILTER");
+
+        rootDescription:CreateTitle(addon.L["Categories"]);
+        CreateCheckbox(rootDescription, addon.L["Merge Small Categories"], addon.Filters.db.profile, {"MergeSmallCategories"});
+
+        rootDescription:CreateDivider();
+        rootDescription:CreateTitle(addon.L["Achievements"]);
+        CreateAchievementFilters(rootDescription, addon.Filters.db.profile);
+
+        local earnedBy = rootDescription:CreateButton(addon.L["Earned By"]);
+        CreateRadio(earnedBy, addon.Filters.Account, addon.Filters.db.profile, {"EarnedBy"}, false);
+        CreateRadio(earnedBy, addon.Filters.CharacterAccount, addon.Filters.db.profile, {"EarnedBy"}, false);
+        CreateRadio(earnedBy, addon.Filters.CharacterOnly, addon.Filters.db.profile, {"EarnedBy"}, false);
+
+        rootDescription:CreateDivider();
+        rootDescription:CreateTitle(addon.L["Tabs"]);
+        for _, t in next, addon.TabsOrder do
+            CreateAchievementFilters(rootDescription:CreateButton(addon.Tabs[t].Text), addon.Tabs[t].Filters);
+        end
+
+        rootDescription:CreateDivider();
+        rootDescription:CreateTitle(addon.L["Special Categories"]);
+        CreateAchievementFilters(rootDescription:CreateButton(addon.L["Current Zone"]), addon.Filters.db.profile.CurrentZone);
+        CreateAchievementFilters(rootDescription:CreateButton(addon.L["Selected Zone"]), addon.Filters.db.profile.SelectedZone);
+        CreateAchievementFilters(rootDescription:CreateButton(addon.L["Tracking Achievements"]), addon.Filters.db.profile.TrackingAchievements);
+
+        rootDescription:CreateDivider();
+        local help = rootDescription:CreateButton((addon.Tutorials.FeaturesTutorial:HasUnviewedPages() and "|T132049:0|t" or "") .. addon.L["Help"]);
+        local pages = addon.Tutorials.FeaturesTutorial.Pages;
+        help:CreateTitle(addon.L["View Tutorial"]);
+        for i, _ in next, pages do
+            help:CreateButton(
+                (pages[i].IsViewed and "" or "|T132049:0|t") .. string.format(addon.Util.Colors.White, addon.Util.Colors.RemoveColor(pages[i].SubTitle)),
+                function()
+                    addon.Tutorials.FeaturesTutorial:ShowTutorial(i);
+                end
+            );
+        end
+
+        rootDescription:CreateButton(addon.L["Options"], addon.Options.Open);
+
+        if addon.Options.db.profile.ShowPlaceholdersFilter then
+            rootDescription:CreateDivider();
+            CreateCheckbox(rootDescription, addon.L["Show placeholders"], addon.Filters.db.profile, {"ShowPlaceholders"});
+        end
+    end);
 end